+++ src/library_pthread.js
@@ -4,8 +4,11 @@
 // found in the LICENSE file.
 
 var LibraryPThread = {
-  $PThread__postset: 'if (!ENVIRONMENT_IS_PTHREAD) PThread.initMainThreadBlock();',
-  $PThread__deps: ['$PROCINFO', '_register_pthread_ptr', 'emscripten_main_thread_process_queued_calls', '$ERRNO_CODES', 'emscripten_futex_wake'],
+  $PThread__postset: 'if (!ENVIRONMENT_IS_PTHREAD) PThread.initMainThreadBlock(); else PThread.initWorker();',
+  $PThread__deps: ['$PROCINFO', '_register_pthread_ptr',
+                   'emscripten_main_thread_process_queued_calls',
+                   '$ERRNO_CODES', 'emscripten_futex_wake', '_kill_thread',
+                   '_cancel_thread', '_cleanup_thread'],
   $PThread: {
     MAIN_THREAD_ID: 1, // A special constant that identifies the main JS thread ID.
     mainThreadInfo: {
@@ -65,6 +68,16 @@ var LibraryPThread = {
       PThread.createProfilerBlock(PThread.mainThreadBlock);
       PThread.setThreadName(PThread.mainThreadBlock, "Browser main thread");
       PThread.setThreadStatus(PThread.mainThreadBlock, {{{ cDefine('EM_THREAD_STATUS_RUNNING') }}});
+#endif
+    },
+    initWorker: function() {
+#if USE_CLOSURE_COMPILER
+      // worker.js is not compiled together with us, and must access certain
+      // things.
+      PThread['receiveObjectTransfer'] = PThread.receiveObjectTransfer;
+      PThread['setThreadStatus'] = PThread.setThreadStatus;
+      PThread['threadCancel'] = PThread.threadCancel;
+      PThread['threadExit'] = PThread.threadExit;
 #endif
     },
     // Maps pthread_t to pthread info objects
@@ -178,7 +191,7 @@ var LibraryPThread = {
         if (ENVIRONMENT_IS_PTHREAD) {
           // Note: in theory we would like to return any offscreen canvases back to the main thread,
           // but if we ever fetched a rendering context for them that would not be valid, so we don't try.
-          postMessage({ cmd: 'exit' });
+          postMessage({ 'cmd': 'exit' });
         }
       }
     },
@@ -190,7 +203,7 @@ var LibraryPThread = {
       _emscripten_futex_wake(threadInfoStruct + {{{ C_STRUCTS.pthread.threadStatus }}}, {{{ cDefine('INT_MAX') }}}); // wake all threads
       threadInfoStruct = selfThreadId = 0; // Not hosting a pthread anymore in this worker, reset the info structures to null.
       __register_pthread_ptr(0, 0, 0); // Unregister the thread block also inside the asm.js scope.
-      postMessage({ cmd: 'cancelDone' });
+      postMessage({ 'cmd': 'cancelDone' });
     },
 
     terminateAllThreads: function() {
@@ -305,32 +318,32 @@ var LibraryPThread = {
 
         // Ask the new worker to load up the Emscripten-compiled page. This is a heavy operation.
         worker.postMessage({
-          cmd: 'load',
+          'cmd': 'load',
           // If the application main .js file was loaded from a Blob, then it is not possible
           // to access the URL of the current script that could be passed to a Web Worker so that
           // it could load up the same file. In that case, developer must either deliver the Blob
           // object in Module['mainScriptUrlOrBlob'], or a URL to it, so that pthread Workers can
           // independently load up the same main application file.
-          urlOrBlob: Module['mainScriptUrlOrBlob'] || _scriptDir,
+          'urlOrBlob': Module['mainScriptUrlOrBlob'] || _scriptDir,
 #if WASM
-          wasmMemory: wasmMemory,
-          wasmModule: wasmModule,
+          'wasmMemory': wasmMemory,
+          'wasmModule': wasmModule,
 #if LOAD_SOURCE_MAP
-          wasmSourceMap: wasmSourceMap,
+          'wasmSourceMap': wasmSourceMap,
 #endif
 #if USE_OFFSET_CONVERTER
-          wasmOffsetConverter: wasmOffsetConverter,
+          'wasmOffsetConverter': wasmOffsetConverter,
 #endif
 #else
-          buffer: HEAPU8.buffer,
-          asmJsUrlOrBlob: Module["asmJsUrlOrBlob"],
+          'buffer': HEAPU8.buffer,
+          'asmJsUrlOrBlob': Module["asmJsUrlOrBlob"],
 #endif
 #if !WASM_BACKEND
-          tempDoublePtr: tempDoublePtr,
+          'tempDoublePtr': tempDoublePtr,
 #endif
-          DYNAMIC_BASE: DYNAMIC_BASE,
-          DYNAMICTOP_PTR: DYNAMICTOP_PTR,
-          PthreadWorkerInit: PthreadWorkerInit
+          'DYNAMIC_BASE': DYNAMIC_BASE,
+          'DYNAMICTOP_PTR': DYNAMICTOP_PTR,
+          'PthreadWorkerInit': PthreadWorkerInit
         });
         PThread.unusedWorkers.push(worker);
       }
@@ -346,34 +359,35 @@ var LibraryPThread = {
         var worker = workers[i];
         (function(worker) {
           worker.onmessage = function(e) {
-            var d = e.data;
+            var d = e['data'];
+            var cmd = d['cmd'];
             // Sometimes we need to backproxy events to the calling thread (e.g. HTML5 DOM events handlers such as emscripten_set_mousemove_callback()), so keep track in a globally accessible variable about the thread that initiated the proxying.
             if (worker.pthread) PThread.currentProxiedOperationCallerThread = worker.pthread.threadInfoStruct;
 
             // If this message is intended to a recipient that is not the main thread, forward it to the target thread.
-            if (d.targetThread && d.targetThread != _pthread_self()) {
+            if (d['targetThread'] && d['targetThread'] != _pthread_self()) {
               var thread = PThread.pthreads[d.targetThread];
               if (thread) {
-                thread.worker.postMessage(e.data, d.transferList);
+                thread.worker.postMessage(e.data, d['transferList']);
               } else {
-                console.error('Internal error! Worker sent a message "' + d.cmd + '" to target pthread ' + d.targetThread + ', but that thread no longer exists!');
+                console.error('Internal error! Worker sent a message "' + cmd + '" to target pthread ' + d['targetThread'] + ', but that thread no longer exists!');
               }
               PThread.currentProxiedOperationCallerThread = undefined;
               return;
             }
 
-            if (d.cmd === 'processQueuedMainThreadWork') {
+            if (cmd === 'processQueuedMainThreadWork') {
               // TODO: Must post message to main Emscripten thread in PROXY_TO_WORKER mode.
               _emscripten_main_thread_process_queued_calls();
-            } else if (d.cmd === 'spawnThread') {
+            } else if (cmd === 'spawnThread') {
               __spawn_thread(e.data);
-            } else if (d.cmd === 'cleanupThread') {
-              __cleanup_thread(d.thread);
-            } else if (d.cmd === 'killThread') {
-              __kill_thread(d.thread);
-            } else if (d.cmd === 'cancelThread') {
-              __cancel_thread(d.thread);
-            } else if (d.cmd === 'loaded') {
+            } else if (cmd === 'cleanupThread') {
+              __cleanup_thread(d['thread']);
+            } else if (cmd === 'killThread') {
+              __kill_thread(d['thread']);
+            } else if (cmd === 'cancelThread') {
+              __cancel_thread(d['thread']);
+            } else if (cmd === 'loaded') {
               worker.loaded = true;
               // If this Worker is already pending to start running a thread, launch the thread now
               if (worker.runPthread) {
@@ -384,34 +398,34 @@ var LibraryPThread = {
               if (numWorkersLoaded === numWorkers && onFinishedLoading) {
                 onFinishedLoading();
               }
-            } else if (d.cmd === 'print') {
-              out('Thread ' + d.threadId + ': ' + d.text);
-            } else if (d.cmd === 'printErr') {
-              err('Thread ' + d.threadId + ': ' + d.text);
-            } else if (d.cmd === 'alert') {
-              alert('Thread ' + d.threadId + ': ' + d.text);
-            } else if (d.cmd === 'exit') {
+            } else if (cmd === 'print') {
+              out('Thread ' + d['threadId'] + ': ' + d['text']);
+            } else if (cmd === 'printErr') {
+              err('Thread ' + d['threadId'] + ': ' + d['text']);
+            } else if (cmd === 'alert') {
+              alert('Thread ' + d['threadId'] + ': ' + d['text']);
+            } else if (cmd === 'exit') {
               var detached = worker.pthread && Atomics.load(HEAPU32, (worker.pthread.thread + {{{ C_STRUCTS.pthread.detached }}}) >> 2);
               if (detached) {
                 PThread.returnWorkerToPool(worker);
               }
-            } else if (d.cmd === 'exitProcess') {
+            } else if (cmd === 'exitProcess') {
               // A pthread has requested to exit the whole application process (runtime).
               noExitRuntime = false;
               try {
-                exit(d.returnCode);
+                exit(d['returnCode']);
               } catch (e) {
                 if (e instanceof ExitStatus) return;
                 throw e;
               }
-            } else if (d.cmd === 'cancelDone') {
+            } else if (cmd === 'cancelDone') {
               PThread.returnWorkerToPool(worker);
-            } else if (d.cmd === 'objectTransfer') {
+            } else if (cmd === 'objectTransfer') {
               PThread.receiveObjectTransfer(e.data);
             } else if (e.data.target === 'setimmediate') {
               worker.postMessage(e.data); // Worker wants to postMessage() to itself to implement setImmediate() emulation.
             } else {
-              err("worker sent an unknown command " + d.cmd);
+              err("worker sent an unknown command " + cmd);
             }
             PThread.currentProxiedOperationCallerThread = undefined;
           };
@@ -483,7 +497,7 @@ var LibraryPThread = {
     if (ENVIRONMENT_IS_PTHREAD) throw 'Internal Error! _cancel_thread() can only ever be called from main application thread!';
     if (!pthread_ptr) throw 'Internal Error! Null pthread_ptr in _cancel_thread!';
     var pthread = PThread.pthreads[pthread_ptr];
-    pthread.worker.postMessage({ cmd: 'cancel' });
+    pthread.worker.postMessage({ 'cmd': 'cancel' });
   },
 
   _spawn_thread: function(threadParams) {
@@ -537,17 +551,17 @@ var LibraryPThread = {
 
     worker.pthread = pthread;
     var msg = {
-        cmd: 'run',
-        start_routine: threadParams.startRoutine,
-        arg: threadParams.arg,
-        threadInfoStruct: threadParams.pthread_ptr,
-        selfThreadId: threadParams.pthread_ptr, // TODO: Remove this since thread ID is now the same as the thread address.
-        parentThreadId: threadParams.parent_pthread_ptr,
-        stackBase: threadParams.stackBase,
-        stackSize: threadParams.stackSize,
+        'cmd': 'run',
+        'start_routine': threadParams.startRoutine,
+        'arg': threadParams.arg,
+        'threadInfoStruct': threadParams.pthread_ptr,
+        'selfThreadId': threadParams.pthread_ptr, // TODO: Remove this since thread ID is now the same as the thread address.
+        'parentThreadId': threadParams.parent_pthread_ptr,
+        'stackBase': threadParams.stackBase,
+        'stackSize': threadParams.stackSize,
 #if OFFSCREENCANVAS_SUPPORT
-        moduleCanvasId: threadParams.moduleCanvasId,
-        offscreenCanvases: threadParams.offscreenCanvases,
+        'moduleCanvasId': threadParams.moduleCanvasId,
+        'offscreenCanvases': threadParams.offscreenCanvases,
 #endif
       };
     worker.runPthread = function() {
@@ -562,7 +576,7 @@ var LibraryPThread = {
   },
 
   _num_logical_cores__deps: ['emscripten_force_num_logical_cores'],
-  _num_logical_cores: '; if (ENVIRONMENT_IS_PTHREAD) __num_logical_cores = PthreadWorkerInit.__num_logical_cores; else { PthreadWorkerInit.__num_logical_cores = __num_logical_cores = {{{ makeStaticAlloc(4) }}}; HEAPU32[__num_logical_cores>>2] = navigator["hardwareConcurrency"] || ' + {{{ PTHREAD_HINT_NUM_CORES }}} + '; }',
+  _num_logical_cores: '; if (ENVIRONMENT_IS_PTHREAD) __num_logical_cores = PthreadWorkerInit["__num_logical_cores"]; else { PthreadWorkerInit["__num_logical_cores"] = __num_logical_cores = {{{ makeStaticAlloc(4) }}}; HEAPU32[__num_logical_cores>>2] = navigator["hardwareConcurrency"] || ' + {{{ PTHREAD_HINT_NUM_CORES }}} + '; }',
 
   emscripten_has_threading_support: function() {
     return typeof SharedArrayBuffer !== 'undefined';
@@ -836,7 +850,7 @@ var LibraryPThread = {
         Atomics.store(HEAPU32, (thread + {{{ C_STRUCTS.pthread.detached }}} ) >> 2, 1); // Mark the thread as detached.
 
         if (!ENVIRONMENT_IS_PTHREAD) __cleanup_thread(thread);
-        else postMessage({ cmd: 'cleanupThread', thread: thread });
+        else postMessage({ 'cmd': 'cleanupThread', 'thread': thread });
         return 0;
       }
       if (!block) {
@@ -881,7 +895,7 @@ var LibraryPThread = {
     }
     if (signal != 0) {
       if (!ENVIRONMENT_IS_PTHREAD) __kill_thread(thread);
-      else postMessage({ cmd: 'killThread', thread: thread});
+      else postMessage({ 'cmd': 'killThread', 'thread': thread});
     }
     return 0;
   },
@@ -903,7 +917,7 @@ var LibraryPThread = {
     }
     Atomics.compareExchange(HEAPU32, (thread + {{{ C_STRUCTS.pthread.threadStatus }}} ) >> 2, 0, 2); // Signal the thread that it needs to cancel itself.
     if (!ENVIRONMENT_IS_PTHREAD) __cancel_thread(thread);
-    else postMessage({ cmd: 'cancelThread', thread: thread});
+    else postMessage({ 'cmd': 'cancelThread', 'thread': thread});
     return 0;
   },
 
@@ -1095,7 +1109,7 @@ var LibraryPThread = {
   },
 
   // Stores the memory address that the main thread is waiting on, if any.
-  _main_thread_futex_wait_address: '; if (ENVIRONMENT_IS_PTHREAD) __main_thread_futex_wait_address = PthreadWorkerInit.__main_thread_futex_wait_address; else PthreadWorkerInit.__main_thread_futex_wait_address = __main_thread_futex_wait_address = {{{ makeStaticAlloc(4) }}}',
+  _main_thread_futex_wait_address: '; if (ENVIRONMENT_IS_PTHREAD) __main_thread_futex_wait_address = PthreadWorkerInit["__main_thread_futex_wait_address"]; else PthreadWorkerInit["__main_thread_futex_wait_address"] = __main_thread_futex_wait_address = {{{ makeStaticAlloc(4) }}}',
 
   // Returns 0 on success, or one of the values -ETIMEDOUT, -EWOULDBLOCK or -EINVAL on error.
   emscripten_futex_wait__deps: ['_main_thread_futex_wait_address', 'emscripten_main_thread_process_queued_calls'],
@@ -1186,7 +1200,7 @@ var LibraryPThread = {
 
   __call_main: function(argc, argv) {
     var returnCode = _main(argc, argv);
-    if (!noExitRuntime) postMessage({ cmd: 'exitProcess', returnCode: returnCode });
+    if (!noExitRuntime) postMessage({ 'cmd': 'exitProcess', 'returnCode': returnCode });
     return returnCode;
   },