+++ tests/test_browser.py
@@ -3597,9 +3597,13 @@ window.close = function() {
     '''))
 
   # Test that the emscripten_ atomics api functions work.
+  @parameterized({
+    'normal': ([],),
+    'closure': (['--closure', '1'],),
+  })
   @requires_threads
-  def test_pthread_atomics(self):
-    self.btest(path_from_root('tests', 'pthread', 'test_pthread_atomics.cpp'), expected='0', args=['-s', 'TOTAL_MEMORY=64MB', '-O3', '-s', 'USE_PTHREADS=1', '-s', 'PTHREAD_POOL_SIZE=8', '-s', 'ALLOW_BLOCKING_ON_MAIN_THREAD'])
+  def test_pthread_atomics(self, args=[]):
+    self.btest(path_from_root('tests', 'pthread', 'test_pthread_atomics.cpp'), expected='0', args=['-s', 'TOTAL_MEMORY=64MB', '-O3', '-s', 'USE_PTHREADS=1', '-s', 'PTHREAD_POOL_SIZE=8', '-g1', '-s', 'ALLOW_BLOCKING_ON_MAIN_THREAD'] + args)
 
   # Test 64-bit atomics.
   @requires_threads